requirements.txt
# Django e dependências principais
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.0
python-decouple>=3.8
pillow>=10.0

# WebSockets e Comunicação em Tempo Real
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Desenvolvimento e ferramentas
django-extensions>=3.2
django-filter>=23.0
dj-database-url>=2.1

# Produção (opcional)
gunicorn>=21.0
whitenoise>=6.5
psycopg2-binary>=2.9  # Para PostgreSQL
